requirements/base.txt
# Base
Django==1.11.11
# pytz==2018.4


# Documentation
mkdocs==0.17.2
# Jinja2==2.10
# Markdown==2.6.9
# MarkupSafe==1.0
# PyYAML==3.12
# click==6.7
# livereload==2.5.1
# six==1.11.0
# tornado==4.5.2


# Django Rest Framework
djangorestframework==3.7.7
django-filter==1.1.0


# Djoser
djoser==0.6.0
# django-templated-mail==1.0.0
